+++ R/counts.R
@@ -23,16 +23,33 @@ NULL
 countcorrect <- function(data, .by = NULL, .input = NULL, .extra = NULL) {
   .input <- list(name_nc = "ncorrect", name_acc = "acc") |>
     update_settings(.input)
+  .extra <- list(
+    name_check = NULL,
+    check_valid = NULL
+  ) |>
+    update_settings(.extra)
+  if (!is.null(.extra$name_check) && !is.null(.extra$check_valid)) {
+    if (!has_name(data, .extra$name_check)) {
+      warn(
+        "Cannot find the specified checking variable. Skip checking.",
+        "miss_chk_var"
+      )
+    } else {
+      data <- data |>
+        filter(.data[[.extra$name_check]] %in% .extra$check_valid)
+    }
+  }
   if (!has_name(data, .input$name_nc)) {
     if (is.character(data[[.input$name_acc]])) {
       # character input uses "-" to separate individual responses
       data <- data |>
         mutate(
           "{.input$name_acc}" := parse_char_resp(
-            .data[[.input$name_acc]]
+            .data[[.input$name_acc]],
+            delim = "\\D"
           )
         ) |>
-        unnest(.data[[.input$name_acc]])
+        unnest(all_of(.input$name_acc))
     }
     data <- data |>
       group_by(across(all_of(.by))) |>